fix(types): fix node16 resolution
Using the tool attw we saw that the default export was not
behaving correctly when node16 or nodenext was used as a resolution
algorithm. We fixed that and intergrated a lot more tests.

We also added support for the legacy node11 (node) resolution algorithm.

BREAKING CHANGES: we changed the exproted types